DTYPE-DEFAULT-001 Phase A+B: Migrate default dtype to float32
## Changes

### Phase A - Inventory & Spec Harmonisation (Complete)
- Catalogued all 37 float64 occurrences across 10 files
- Identified gradient-critical paths requiring float64
- Proposed documentation updates for arch.md and runtime checklist
- Risk assessment complete (low-medium risk, high performance benefit)

### Phase B - Implementation Update (Substantially Complete)
Core Defaults (B1):
- CLI default: float64 → float32 (__main__.py:365)
- Crystal/Detector/Simulator constructors: float64 → float32 defaults
- HKL I/O functions: float64 → float32 defaults (Fdump binary preserved at float64)

Constants & Initialization (B2):
- Moved dtype/device resolution early in __main__.py (line 793)
- Made beam_direction/polarization_axis tensors dtype-aware
- Made HKL data tensors dtype-aware
- Updated auto_selection.py to accept dtype parameter
- All tensor creations now use resolved dtype variable

Remaining (B3 - deferred):
- io/source.py: 6 occurrences still hard-coded (low priority, rare usage)
- utils/noise.py, utils/c_random.py: 1 each (low priority)

## Test Status
✅ CLI help works (smoke test passed)
❌ AT-PARALLEL-012: 43/50 peaks (regression from previous pass)
   - Correlation still high (≥0.9995 assertion passed)
   - Peak matching affected by native float32 vs float64→float32 cast
   - Needs investigation in Phase C

## Artifacts
- reports/DTYPE-DEFAULT-001/inventory.md (37 items catalogued)
- reports/DTYPE-DEFAULT-001/proposed_doc_changes.md (6 docs to update)
- reports/DTYPE-DEFAULT-001/phase_b_summary.md (implementation summary)

## Module Scope
data models, I/O, CLI/config

## Next Actions
1. Debug AT-PARALLEL-012 peak matching regression
2. Complete remaining B3 items (io/source.py, utils/noise.py)
3. Run full AT-PARALLEL test suite (Phase C)
4. Update documentation (Phase D)

Test suite: partial (1 failed AT-012, 0 passed)

+++ src/nanobrag_torch/io/hkl.py
@@ -16,7 +16,7 @@ def read_hkl_file(
     filepath: str,
     default_F: float = 0.0,
     device=None,
-    dtype=torch.float64
+    dtype=torch.float32
 ) -> Tuple[torch.Tensor, dict]:
     """
     Read HKL text file with two-pass algorithm matching C implementation.
@@ -212,7 +212,7 @@ def try_load_hkl_or_fdump(
     default_F: float = 0.0,
     write_cache: bool = True,
     device=None,
-    dtype=torch.float64
+    dtype=torch.float32
 ) -> Tuple[Optional[torch.Tensor], Optional[dict]]:
     """
     Attempt to load HKL data with caching behavior matching C code.
